requirements.txt
fastapi>=0.110.0,<1.0.0
uvicorn[standard]>=0.27.0,<1.0.0
pillow>=10.0.0,<11.0.0
pydantic>=2.0.0,<3.0.0
pydantic-settings>=2.0.0,<3.0.0
pytest>=8.0.0,<9.0.0
httpx>=0.25.0,<0.28.0
python-multipart>=0.0.9,<0.0.10
